possible fix for borken turnspeed controller

+++ selfdrive/controls/lib/turn_speed_controller.py
@@ -118,7 +118,7 @@ class TurnSpeedController():
     distance_since_fix = self._v_ego * gps_fix_age
     distances_to_sections_ahead = np.maximum(0., np.array(map_data.turnSpeedLimitsAheadDistances) - distance_since_fix)
     speed_limit_in_sections_ahead = map_data.turnSpeedLimitsAhead
-    turn_signs_in_sections_ahead = map_data.turnSpeedLimitsAheadSigns * 1.4
+    turn_signs_in_sections_ahead = int(map_data.turnSpeedLimitsAheadSigns * 1.4)
 
     # Ensure current speed limit is considered only if we are inside the section.
     if map_data.turnSpeedLimitValid and self._v_ego > 0.:
